requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers>=4.41.2
pillow>=9.4.0
opencv-python>=4.7.0
numpy>=1.23.5
tqdm>=4.65.0
huggingface-hub>=0.16.4
